# contracts/openapi.yaml
openapi: 3.1.0
info:
  title: 电气车间备件管理系统
  version: 1.0.0
paths:
...
  /api/v1/purchase-requests:
    get:
      tags:
      - 请购
      summary: Requests
      operationId: requests_api_v1_purchase_requests_get
      security:
      - HTTPBearer: []
      parameters:
      - name: page
        in: query
        required: false
        schema:
          type: integer
          minimum: 1
          default: 1
          title: Page
      - name: page_size
        in: query
        required: false
        schema:
          type: integer
          maximum: 200
          minimum: 1
          default: 20
          title: Page Size
      - name: status
        in: query
        required: false
        schema:
          anyOf:
          - $ref: '#/components/schemas/PurchaseRequestStatus'
          - type: 'null'
          title: Status
      - name: keyword
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Keyword
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Page_PurchaseRequestRead_'
        '400':
          description: 业务校验失败
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '401':
          description: 未认证或凭证无效
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '403':
          description: 权限不足
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '404':
          description: 资源不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '409':
          description: 版本、状态或业务数据冲突
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '422':
          description: 请求参数校验失败
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
    post:
      tags:
      - 请购
      summary: Create
      operationId: create_api_v1_purchase_requests_post
      security:
      - HTTPBearer: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PurchaseRequestCreate'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PurchaseRequestRead'
        '400':
          description: 业务校验失败
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '401':
          description: 未认证或凭证无效
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '403':
          description: 权限不足
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '404':
          description: 资源不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '409':
          description: 版本、状态或业务数据冲突
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '422':
          description: 请求参数校验失败
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
  /api/v1/purchase-requests/{request_id}:
    get:
      tags:
      - 请购
      summary: Detail
      operationId: detail_api_v1_purchase_requests__request_id__get
      security:
      - HTTPBearer: []
      parameters:
      - name: request_id
        in: path
        required: true
        schema:
          type: integer
          title: Request Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PurchaseRequestRead'
        '400':
          description: 业务校验失败
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '401':
          description: 未认证或凭证无效
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '403':
          description: 权限不足
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '404':
          description: 资源不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '409':
          description: 版本、状态或业务数据冲突
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '422':
          description: 请求参数校验失败
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
    patch:
      tags:
      - 请购
      summary: Update
      operationId: update_api_v1_purchase_requests__request_id__patch
      security:
      - HTTPBearer: []
      parameters:
      - name: request_id
        in: path
        required: true
        schema:
          type: integer
          title: Request Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PurchaseRequestUpdate'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PurchaseRequestRead'
        '400':
          description: 业务校验失败
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '401':
          description: 未认证或凭证无效
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '403':
          description: 权限不足
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '404':
          description: 资源不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '409':
          description: 版本、状态或业务数据冲突
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '422':
          description: 请求参数校验失败
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
  /api/v1/purchase-requests/{request_id}/submit:
    post:
      tags:
      - 请购
      summary: Submit
      operationId: submit_api_v1_purchase_requests__request_id__submit_post
      security:
      - HTTPBearer: []
      parameters:
      - name: request_id
        in: path
        required: true
        schema:
          type: integer
          title: Request Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ActionVersion'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PurchaseRequestRead'
        '400':
          description: 业务校验失败
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '401':
          description: 未认证或凭证无效
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '403':
          description: 权限不足
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '404':
          description: 资源不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '409':
          description: 版本、状态或业务数据冲突
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '422':
          description: 请求参数校验失败
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
  /api/v1/purchase-requests/{request_id}/accept:
    post:
      tags:
      - 请购
      summary: Accept
      operationId: accept_api_v1_purchase_requests__request_id__accept_post
      security:
      - HTTPBearer: []
      parameters:
      - name: request_id
        in: path
        required: true
        schema:
          type: integer
          title: Request Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ActionVersion'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PurchaseRequestRead'
        '400':
          description: 业务校验失败
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '401':
          description: 未认证或凭证无效
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '403':
          description: 权限不足
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '404':
          description: 资源不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '409':
          description: 版本、状态或业务数据冲突
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '422':
          description: 请求参数校验失败
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
  /api/v1/purchase-requests/{request_id}/return:
    post:
      tags:
      - 请购
      summary: Return Item
      operationId: return_item_api_v1_purchase_requests__request_id__return_post
      security:
      - HTTPBearer: []
      parameters:
      - name: request_id
        in: path
        required: true
        schema:
          type: integer
          title: Request Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReasonAction'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PurchaseRequestRead'
        '400':
          description: 业务校验失败
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '401':
          description: 未认证或凭证无效
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '403':
          description: 权限不足
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '404':
          description: 资源不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '409':
          description: 版本、状态或业务数据冲突
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '422':
          description: 请求参数校验失败
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
  /api/v1/purchase-requests/{request_id}/cancel:
    post:
      tags:
      - 请购
      summary: Cancel
      operationId: cancel_api_v1_purchase_requests__request_id__cancel_post
      security:
      - HTTPBearer: []
      parameters:
      - name: request_id
        in: path
        required: true
        schema:
          type: integer
          title: Request Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ActionVersion'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PurchaseRequestRead'
        '400':
          description: 业务校验失败
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '401':
          description: 未认证或凭证无效
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '403':
          description: 权限不足
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '404':
          description: 资源不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '409':
          description: 版本、状态或业务数据冲突
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '422':
          description: 请求参数校验失败
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
  /api/v1/purchase-requests/{request_id}/close:
    post:
      tags:
      - 请购
      summary: Close
      operationId: close_api_v1_purchase_requests__request_id__close_post
      security:
      - HTTPBearer: []
      parameters:
      - name: request_id
        in: path
        required: true
        schema:
          type: integer
          title: Request Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReasonAction'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PurchaseRequestRead'
        '400':
          description: 业务校验失败
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '401':
          description: 未认证或凭证无效
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '403':
          description: 权限不足
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '404':
          description: 资源不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '409':
          description: 版本、状态或业务数据冲突
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '422':
          description: 请求参数校验失败
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
  /api/v1/purchase-requests/{request_id}/receipts:
    get:
      tags:
      - 请购
      summary: Receipts
      operationId: receipts_api_v1_purchase_requests__request_id__receipts_get
      security:
      - HTTPBearer: []
      parameters:
      - name: request_id
        in: path
        required: true
        schema:
          type: integer
          title: Request Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/StockOperationRead'
                title: Response Receipts Api V1 Purchase Requests  Request Id  Receipts
                  Get
        '400':
          description: 业务校验失败
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '401':
          description: 未认证或凭证无效
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '403':
          description: 权限不足
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '404':
          description: 资源不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '409':
          description: 版本、状态或业务数据冲突
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '422':
          description: 请求参数校验失败
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
  /api/v1/purchase-request-lines/{line_id}/prepare-inbound:
    post:
      tags:
      - 请购
      summary: Prepare Inbound
      operationId: prepare_inbound_api_v1_purchase_request_lines__line_id__prepare_inbound_post
      security:
      - HTTPBearer: []
      parameters:
      - name: line_id
        in: path
        required: true
        schema:
          type: integer
          title: Line Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PreparedInboundRead'
        '400':
          description: 业务校验失败
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '401':
          description: 未认证或凭证无效
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '403':
          description: 权限不足
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '404':
          description: 资源不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '409':
          description: 版本、状态或业务数据冲突
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '422':
          description: 请求参数校验失败
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
...
components:
  schemas:
    ActionVersion:
      properties:
        version:
          anyOf:
          - type: integer
          - type: 'null'
          title: Version
      additionalProperties: false
      type: object
      title: ActionVersion
    ApiError:
      properties:
        code:
          type: string
          title: Code
        message:
          type: string
          title: Message
        details:
          additionalProperties: true
          type: object
          title: Details
        request_id:
          type: string
          title: Request Id
      type: object
      required:
      - code
      - message
      - request_id
      title: ApiError
...
    BusinessEventRead:
      properties:
        id:
          type: integer
          title: Id
        action:
          type: string
          title: Action
        old_status:
          anyOf:
          - type: string
          - type: 'null'
          title: Old Status
        new_status:
          anyOf:
          - type: string
          - type: 'null'
          title: New Status
        operator_name:
          type: string
          title: Operator Name
        occurred_at:
          type: string
          format: date-time
          title: Occurred At
        remark:
          anyOf:
          - type: string
          - type: 'null'
          title: Remark
      type: object
      required:
      - id
      - action
      - operator_name
      - occurred_at
      title: BusinessEventRead
...
    OperationType:
      type: string
      enum:
      - INBOUND
      - OUTBOUND
      title: OperationType
...
    Page_PurchaseRequestRead_:
      properties:
        items:
          items:
            $ref: '#/components/schemas/PurchaseRequestRead'
          type: array
          title: Items
        page:
          type: integer
          title: Page
        page_size:
          type: integer
          title: Page Size
        total:
          type: integer
          title: Total
      type: object
      required:
      - items
      - page
      - page_size
      - total
      title: Page[PurchaseRequestRead]
...
    PreparedInboundRead:
      properties:
        purchase_request_no:
          type: string
          title: Purchase Request No
        line_id:
          type: integer
          title: Line Id
        purchase_material_id:
          type: integer
          title: Purchase Material Id
        material_name:
          type: string
          title: Material Name
        model_spec:
          type: string
          title: Model Spec
        unit_name:
          type: string
          title: Unit Name
        remaining_qty:
          type: string
          pattern: ^(?!^[-+.]*$)[+-]?0*\d*\.?\d*$
          title: Remaining Qty
        stock_material_id:
          anyOf:
          - type: integer
          - type: 'null'
          title: Stock Material Id
      type: object
      required:
      - purchase_request_no
      - line_id
      - purchase_material_id
      - material_name
      - model_spec
      - unit_name
      - remaining_qty
      title: PreparedInboundRead
...
    PurchaseRequestCreate:
      properties:
        request_no:
          anyOf:
          - type: string
            maxLength: 128
            minLength: 1
          - type: 'null'
          title: Request No
        remark:
          anyOf:
          - type: string
            maxLength: 1000
          - type: 'null'
          title: Remark
        lines:
          items:
            $ref: '#/components/schemas/PurchaseRequestLineWrite'
          type: array
          title: Lines
      additionalProperties: false
      type: object
      title: PurchaseRequestCreate
    PurchaseRequestLineRead:
      properties:
        id:
          type: integer
          title: Id
        purchase_material_id:
          type: integer
          title: Purchase Material Id
        material_code_snapshot:
          anyOf:
          - type: string
          - type: 'null'
          title: Material Code Snapshot
        material_name_snapshot:
          type: string
          title: Material Name Snapshot
        model_spec_snapshot:
          type: string
          title: Model Spec Snapshot
        unit_name_snapshot:
          type: string
          title: Unit Name Snapshot
        requested_qty:
          type: string
          pattern: ^(?!^[-+.]*$)[+-]?0*\d*\.?\d*$
          title: Requested Qty
        received_qty:
          type: string
          pattern: ^(?!^[-+.]*$)[+-]?0*\d*\.?\d*$
          title: Received Qty
        usage:
          type: string
          title: Usage
        project_subitem_id:
          type: integer
          title: Project Subitem Id
        project_code_snapshot:
          type: string
          title: Project Code Snapshot
        subitem_no_snapshot:
          type: string
          title: Subitem No Snapshot
        subitem_name_snapshot:
          type: string
          title: Subitem Name Snapshot
      type: object
      required:
      - id
      - purchase_material_id
      - material_name_snapshot
      - model_spec_snapshot
      - unit_name_snapshot
      - requested_qty
      - received_qty
      - usage
      - project_subitem_id
      - project_code_snapshot
      - subitem_no_snapshot
      - subitem_name_snapshot
      title: PurchaseRequestLineRead
    PurchaseRequestLineWrite:
      properties:
        id:
          anyOf:
          - type: integer
          - type: 'null'
          title: Id
        purchase_material_id:
          type: integer
          title: Purchase Material Id
        requested_qty:
          anyOf:
          - type: number
            exclusiveMinimum: 0.0
          - type: string
            pattern: ^(?!^[-+.]*$)[+-]?0*(?:\d{0,15}|(?=[\d.]{1,19}0*$)\d{0,15}\.\d{0,3}0*$)
          title: Requested Qty
        usage:
          type: string
          maxLength: 500
          minLength: 1
          title: Usage
        project_subitem_id:
          type: integer
          title: Project Subitem Id
      additionalProperties: false
      type: object
      required:
      - purchase_material_id
      - requested_qty
      - usage
      - project_subitem_id
      title: PurchaseRequestLineWrite
    PurchaseRequestRead:
      properties:
        id:
          type: integer
          title: Id
        request_no:
          type: string
          title: Request No
        status:
          $ref: '#/components/schemas/PurchaseRequestStatus'
        applicant_name:
          type: string
          title: Applicant Name
        handler_name:
          anyOf:
          - type: string
          - type: 'null'
          title: Handler Name
        remark:
          anyOf:
          - type: string
          - type: 'null'
          title: Remark
        return_reason:
          anyOf:
          - type: string
          - type: 'null'
          title: Return Reason
        close_reason:
          anyOf:
          - type: string
          - type: 'null'
          title: Close Reason
        submitted_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Submitted At
        completed_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Completed At
        created_at:
          type: string
          format: date-time
          title: Created At
        version:
          type: integer
          title: Version
        lines:
          items:
            $ref: '#/components/schemas/PurchaseRequestLineRead'
          type: array
          title: Lines
        events:
          items:
            $ref: '#/components/schemas/BusinessEventRead'
          type: array
          title: Events
      type: object
      required:
      - id
      - request_no
      - status
      - applicant_name
      - created_at
      - version
      - lines
      - events
      title: PurchaseRequestRead
    PurchaseRequestStatus:
      type: string
      enum:
      - DRAFT
      - SUBMITTED
      - PROCESSING
      - RETURNED
      - PARTIALLY_RECEIVED
      - COMPLETED
      - CLOSED
      - CANCELED
      title: PurchaseRequestStatus
    PurchaseRequestUpdate:
      properties:
        request_no:
          anyOf:
          - type: string
            maxLength: 128
            minLength: 1
          - type: 'null'
          title: Request No
        remark:
          anyOf:
          - type: string
            maxLength: 1000
          - type: 'null'
          title: Remark
        lines:
          items:
            $ref: '#/components/schemas/PurchaseRequestLineWrite'
          type: array
          title: Lines
        version:
          type: integer
          title: Version
      additionalProperties: false
      type: object
      required:
      - version
      title: PurchaseRequestUpdate
    ReasonAction:
      properties:
        version:
          anyOf:
          - type: integer
          - type: 'null'
          title: Version
        reason:
          type: string
          maxLength: 500
          minLength: 1
          title: Reason
      additionalProperties: false
      type: object
      required:
      - reason
      title: ReasonAction
...
    SourceType:
      type: string
      enum:
      - MANUAL
      - PURCHASE_RECEIPT
      - REVERSAL
      - INITIALIZATION
      title: SourceType
...
    StockOperationRead:
      properties:
        id:
          type: integer
          title: Id
        operation_no:
          type: string
          title: Operation No
        operation_type:
          $ref: '#/components/schemas/OperationType'
        occurred_at:
          type: string
          format: date-time
          title: Occurred At
        operator_name:
          type: string
          title: Operator Name
        business_reason:
          type: string
          title: Business Reason
        receiver_name:
          anyOf:
          - type: string
          - type: 'null'
          title: Receiver Name
        project_subitem_id:
          anyOf:
          - type: integer
          - type: 'null'
          title: Project Subitem Id
        source_type:
          $ref: '#/components/schemas/SourceType'
        reversal_of_id:
          anyOf:
          - type: integer
          - type: 'null'
          title: Reversal Of Id
        purchase_request_no:
          anyOf:
          - type: string
          - type: 'null'
          title: Purchase Request No
        client_request_id:
          type: string
          title: Client Request Id
        lines:
          items:
            $ref: '#/components/schemas/StockOperationLineRead'
          type: array
          title: Lines
        created_at:
          type: string
          format: date-time
          title: Created At
        version:
          type: integer
          title: Version
      type: object
      required:
      - id
      - operation_no
      - operation_type
      - occurred_at
      - operator_name
      - business_reason
      - source_type
      - client_request_id
      - lines
      - created_at
      - version
      title: StockOperationRead
...
  securitySchemes:
    HTTPBearer:
      type: http
      scheme: bearer